feat: Add react-router-dom and sweetalert2 dependencies
This commit adds the react-router-dom and sweetalert2 dependencies to the project. These dependencies are required for implementing routing functionality and displaying pop-up alerts in the application.

The react-router-dom package is added to the "dependencies" section of the package.json file with version "^6.27.0". This package provides the necessary components and utilities for handling routing in a React application.

The sweetalert2 package is also added to the "dependencies" section of the package.json file with version "^11.14.4". This package is used for displaying attractive and customizable pop-up alerts in the application.

These dependencies will enhance the functionality and user experience of the application.